Dijkstra's Algorithm
Dijkstra's Algorithm is a graph search algorithm that solves the
single-source shortest path problem for a graph with non-negative edge
path costs, producing a shortest path tree. This algorithm is often used
in routing and as a subroutine in other graph algorithms. This is a C++
implementation of the algorithm.